🔨 refactor: 중복 ValidateGraph 컴포넌트 재사용
